backend/requirements.txt
fastapi[standard]
uvicorn
websockets
python-dotenv
google-generativeai
pydantic
